+++ frontend/src/components/Layout/PageLayout.tsx
@@ -1,7 +1,7 @@
 import { useBooks } from "components/Book";
 import Pagination from "components/Pagination";
 import { Label } from "pages/CommonPages.styled";
-import React, { ReactNode, useEffect, useState } from "react";
+import React, { ReactNode } from "react";
 import { useSearchParams } from "react-router-dom";
 import { StyledCommonWrapper } from "styles/CommonStyled";
 import { IBookItem } from "types/Books";
@@ -14,35 +14,12 @@ interface IPageLayoutProps {
 }
 
 const PageLayout: React.FC<IPageLayoutProps> = ({ children, label, books }) => {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const { currentPage, setCurrentPage, totalPages } = useBooks();
-
-  const handlePageChange = (page: number) => {
-    if (page !== currentPage) {
-      setCurrentPage(page);
-    }
-    const keyword = searchParams.get("keyword") || "";
-
-    setSearchParams({
-      keyword,
-      page: page.toString(),
-    });
-  };
-
   return (
     <StyledCommonWrapper>
       <PageLayoutWrapper>
         {label && <Label>{label}</Label>}
 
         {children}
-
-        {totalPages > 1 && (
-          <Pagination
-            currentPage={currentPage}
-            totalPages={totalPages}
-            onPageChange={handlePageChange}
-          />
-        )}
       </PageLayoutWrapper>
     </StyledCommonWrapper>
   );
